+++ edge-mqtt/src/lib.rs
@@ -1,5 +1,0 @@
-#![warn(clippy::large_futures)]
-#![allow(clippy::uninlined_format_args)]
-#![allow(unknown_lints)]
-
-pub mod io;