+++ edge-raw/src/io.rs
@@ -62,8 +62,7 @@ where
     }
 }
 
-#[cfg(feature = "std")]
-impl<E> std::error::Error for Error<E> where E: std::error::Error {}
+impl<E> core::error::Error for Error<E> where E: core::error::Error {}
 
 /// A utility struct allowing to send and receive UDP packets over a raw socket.
 ///