+++ edge-captive/CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2026-01-01
+* Update the `edge-nal` dependency
+
 ## [0.6.0] - 2025-05-29
 * Optional `defmt` support via two new features (one has to specify one, or the other, or neither, but not both):
   * `log` - uses the `log` crate for all logging