+++ edge-nal-embassy/CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 2026-01-01
+* Update the `edge-nal` and `embedded-io-async` dependencies
+
 ## [0.7.0] - 2025-11-23
 * Erase the const generics from the edge-nal-embassy types #78
 * Make `Tcp`, `Udp` and `Dns` `Copy`